// src/generated/types/authUser.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */

/**
 * Authentication user data
 */
export interface AuthUser {
  lastName?: string;
  firstName?: string;
  username?: string;
  email?: string;
  telephone?: string;
  status?: string;
  created?: number;
  lastAccess?: number;
  disabled?: boolean;
}